combine gpu multiprocess tests and use pytest with `forked` option to run each test in a separate subprocess to WAR multiple `jax.distributed.init` within a single test file/process
